requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.2
# Data Extraction
yfinance==0.2.33
requests==2.31.0
beautifulsoup4==4.12.2
feedparser==6.0.10
# Database
psycopg2-binary==2.9.9
sqlalchemy==1.4.53
# Airflow
apache-airflow==2.7.3
apache-airflow-providers-postgres==5.7.1
# Technical Analysis
ta==0.11.0
# Sentiment Analysis
textblob==0.17.1
vaderSentiment==3.3.2
# API Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
# Dashboard
streamlit==1.29.0
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0
# Notifications
python-telegram-bot==20.7
# smtplib is built-in, no need to install
# Testing
pytest==7.4.3
pytest-cov==4.1.0
# Utilities
schedule==1.2.0
pytz==2023.3